requirements.txt
pymobiledevice3==2.43.2
PyYAML==6.0.1
geopy==2.4.1
